use 64 bit integers for coordinates (and tables)

+++ common.h
@@ -132,8 +132,8 @@ cube apply_moves(cube, int *, int);
 //
 int get_co(cube);
 int get_eo(cube);
-void set_co(cube *, int);
-void set_eo(cube *, int);
+void set_co(cube *, long long);
+void set_eo(cube *, long long);
 int get_tetrad_twist(cube);
 void set_tetrad_twist(cube *, int);
 
@@ -157,18 +157,18 @@ void gui_wait_for_close();
 // table.c
 typedef struct
 {
-    int size; // size in bytes
+    long long size; // size in bytes
     int bits; // bits per entry
     int mask; // (1<<bits)-1
     int divisor; // sizeof(data[0])/bits
     char filename[256];
     unsigned *data;
 } table;
-table table_new(int size, int bits, char *filename);
+table table_new(long long size, int bits, char *filename);
 int table_read(table t);
 int table_write(table t);
-void table_set(table t, int i, int x);
-int table_get(table t, int i);
+void table_set(table t, long long i, int x);
+int table_get(table t, long long i);
 
 //
 table init_tetrad_twist_table(void);
